requirements.txt
notebook
jupyter_contrib_nbextensions
jupyter_nbextensions_configurator
pytorch
torchvision
fastprogress
numpy
pandas
matplotlib
seaborn
nltk
wordcloud
ax-platform
wandb